create ima generic for reuse similar code
